fix(gltf): resolve Basis formats before worker parsing
