+++ modules/gltf/test/gltf-loader.spec.ts
@@ -3,9 +3,10 @@ import test from 'tape-promise/tape';
 import {validateLoader} from 'test/common/conformance';
 
 import {registerLoaders, load, parseSync, fetchFile} from '@loaders.gl/core';
-import {GLTFLoader, postProcessGLTF} from '@loaders.gl/gltf';
+import {GLTFLoader, postProcessGLTF, type GLTFLoaderOptions} from '@loaders.gl/gltf';
 import {DracoLoader} from '@loaders.gl/draco';
 import {ImageBitmapLoader} from '@loaders.gl/images';
+import {getGLTFImageOptions} from '../src/lib/parsers/parse-gltf';
 
 const GLTF_BINARY_URL = '@loaders.gl/gltf/test/data/gltf-2.0/2CylinderEngine.glb';
 const GLTF_JSON_URL = '@loaders.gl/gltf/test/data/gltf-2.0/2CylinderEngine.gltf';
@@ -50,6 +51,48 @@ test('GLTFLoader#load(binary)', async t => {
 test('GLTFLoader#load(text)', async t => {
   const data = await load(GLTF_JSON_URL, GLTFLoader, {gltf: {loadImages: false}});
   t.ok(data.json.asset, 'GLTFLoader returned parsed data');
+  t.end();
+});
+
+test('GLTFLoader#Basis image options', t => {
+  const loaderOptions: GLTFLoaderOptions = {
+    core: {worker: true},
+    basis: {
+      supportedTextureFormats: ['bc3-rgba-unorm'],
+      containerFormat: 'ktx2'
+    }
+  };
+
+  const imageOptions = getGLTFImageOptions(loaderOptions, 'image/ktx2');
+
+  t.deepEqual(
+    imageOptions.basis,
+    {
+      supportedTextureFormats: ['bc3-rgba-unorm'],
+      containerFormat: 'ktx2',
+      format: {alpha: 'bc3', noAlpha: 'bc1'}
+    },
+    'selects a concrete format while preserving partial Basis options'
+  );
+  t.equal(imageOptions.core?.worker, true, 'preserves core options');
+  t.equal(imageOptions.core?.mimeType, 'image/ktx2', 'sets the image MIME type');
+  t.notOk(loaderOptions.basis?.format, 'does not mutate the supplied options');
+
+  const explicitFormatOptions = getGLTFImageOptions({
+    basis: {
+      format: 'rgba32',
+      module: 'encoder'
+    }
+  });
+
+  t.deepEqual(
+    explicitFormatOptions.basis,
+    {
+      format: 'rgba32',
+      module: 'encoder'
+    },
+    'preserves an explicit Basis format and other options'
+  );
   t.end();
 });
