+++ modules/textures/src/lib/parsers/parse-basis.ts
@@ -25,30 +25,11 @@ import {
   GL_RGBA4,
   GL_RGBA8
 } from '../gl-extensions';
-import {detectSupportedTextureFormats} from '../utils/detect-supported-texture-formats';
+import {selectSupportedBasisFormat, type BasisFormat} from '../utils/basis-format-utils';
 import {isKTX} from './parse-ktx';
 
 // TODO - circular type import
 import type {BasisLoaderOptions} from '../../basis-loader';
-
-export type BasisFormat =
-  | 'etc1'
-  | 'etc2'
-  | 'bc1'
-  | 'bc3'
-  | 'bc4'
-  | 'bc5'
-  | 'bc7-m6-opaque-only'
-  | 'bc7-m5'
-  | 'pvrtc1-4-rgb'
-  | 'pvrtc1-4-rgba'
-  | 'astc-4x4'
-  | 'atc-rgb'
-  | 'atc-rgba-interpolated-alpha'
-  | 'rgba32'
-  | 'rgb565'
-  | 'bgr565'
-  | 'rgba4444';
 
 type BasisOutputOptions = {
   basisFormat: number;
@@ -430,174 +411,4 @@ function getBasisOptions(options: BasisLoaderOptions, hasAlpha: boolean): BasisO
     throw new Error(`Unknown Basis format ${format}`);
   }
   return basisOutputOptions;
-}
-
-/**
- * Select transcode format from the list of supported formats
- * @returns key for OutputFormat map
- */
-export function selectSupportedBasisFormat():
-  | BasisFormat
-  | {
-      alpha: BasisFormat;
-      noAlpha: BasisFormat;
-    };
-export function selectSupportedBasisFormat(supportedTextureFormats?: Iterable<TextureFormat>):
-  | BasisFormat
-  | {
-      alpha: BasisFormat;
-      noAlpha: BasisFormat;
-    };
-export function selectSupportedBasisFormat(
-  supportedTextureFormats: Iterable<TextureFormat> = detectSupportedTextureFormats()
-): BasisFormat | {alpha: BasisFormat; noAlpha: BasisFormat} {
-  const textureFormats = new Set(supportedTextureFormats);
-
-  if (hasSupportedTextureFormat(textureFormats, ['astc-4x4-unorm', 'astc-4x4-unorm-srgb'])) {
-    return 'astc-4x4';
-  } else if (hasSupportedTextureFormat(textureFormats, ['bc7-rgba-unorm', 'bc7-rgba-unorm-srgb'])) {
-    return {
-      alpha: 'bc7-m5',
-      noAlpha: 'bc7-m6-opaque-only'
-    };
-  } else if (
-    hasSupportedTextureFormat(textureFormats, [
-      'bc1-rgb-unorm-webgl',
-      'bc1-rgb-unorm-srgb-webgl',
-      'bc1-rgba-unorm',
-      'bc1-rgba-unorm-srgb',
-      'bc2-rgba-unorm',
-      'bc2-rgba-unorm-srgb',
-      'bc3-rgba-unorm',
-      'bc3-rgba-unorm-srgb'
-    ])
-  ) {
-    return {
-      alpha: 'bc3',
-      noAlpha: 'bc1'
-    };
-  } else if (
-    hasSupportedTextureFormat(textureFormats, [
-      'pvrtc-rgb4unorm-webgl',
-      'pvrtc-rgba4unorm-webgl',
-      'pvrtc-rgb2unorm-webgl',
-      'pvrtc-rgba2unorm-webgl'
-    ])
-  ) {
-    return {
-      alpha: 'pvrtc1-4-rgba',
-      noAlpha: 'pvrtc1-4-rgb'
-    };
-  } else if (
-    hasSupportedTextureFormat(textureFormats, [
-      'etc2-rgb8unorm',
-      'etc2-rgb8unorm-srgb',
-      'etc2-rgb8a1unorm',
-      'etc2-rgb8a1unorm-srgb',
-      'etc2-rgba8unorm',
-      'etc2-rgba8unorm-srgb',
-      'eac-r11unorm',
-      'eac-r11snorm',
-      'eac-rg11unorm',
-      'eac-rg11snorm'
-    ])
-  ) {
-    return 'etc2';
-  } else if (textureFormats.has('etc1-rgb-unorm-webgl')) {
-    return 'etc1';
-  } else if (
-    hasSupportedTextureFormat(textureFormats, [
-      'atc-rgb-unorm-webgl',
-      'atc-rgba-unorm-webgl',
-      'atc-rgbai-unorm-webgl'
-    ])
-  ) {
-    return {
-      alpha: 'atc-rgba-interpolated-alpha',
-      noAlpha: 'atc-rgb'
-    };
-  }
-  return 'rgb565';
-}
-
-export function getSupportedBasisFormats(
-  supportedTextureFormats: Iterable<TextureFormat> = detectSupportedTextureFormats()
-): BasisFormat[] {
-  const textureFormats = new Set(supportedTextureFormats);
-  const basisFormats: BasisFormat[] = [];
-
-  if (hasSupportedTextureFormat(textureFormats, ['astc-4x4-unorm', 'astc-4x4-unorm-srgb'])) {
-    basisFormats.push('astc-4x4');
-  }
-  if (
-    hasSupportedTextureFormat(textureFormats, [
-      'bc1-rgb-unorm-webgl',
-      'bc1-rgb-unorm-srgb-webgl',
-      'bc1-rgba-unorm',
-      'bc1-rgba-unorm-srgb',
-      'bc2-rgba-unorm',
-      'bc2-rgba-unorm-srgb',
-      'bc3-rgba-unorm',
-      'bc3-rgba-unorm-srgb'
-    ])
-  ) {
-    basisFormats.push('bc1', 'bc3');
-  }
-  if (hasSupportedTextureFormat(textureFormats, ['bc4-r-unorm', 'bc4-r-snorm'])) {
-    basisFormats.push('bc4');
-  }
-  if (hasSupportedTextureFormat(textureFormats, ['bc5-rg-unorm', 'bc5-rg-snorm'])) {
-    basisFormats.push('bc5');
-  }
-  if (hasSupportedTextureFormat(textureFormats, ['bc7-rgba-unorm', 'bc7-rgba-unorm-srgb'])) {
-    basisFormats.push('bc7-m5', 'bc7-m6-opaque-only');
-  }
-  if (
-    hasSupportedTextureFormat(textureFormats, [
-      'pvrtc-rgb4unorm-webgl',
-      'pvrtc-rgba4unorm-webgl',
-      'pvrtc-rgb2unorm-webgl',
-      'pvrtc-rgba2unorm-webgl'
-    ])
-  ) {
-    basisFormats.push('pvrtc1-4-rgb', 'pvrtc1-4-rgba');
-  }
-  if (
-    hasSupportedTextureFormat(textureFormats, [
-      'etc2-rgb8unorm',
-      'etc2-rgb8unorm-srgb',
-      'etc2-rgb8a1unorm',
-      'etc2-rgb8a1unorm-srgb',
-      'etc2-rgba8unorm',
-      'etc2-rgba8unorm-srgb',
-      'eac-r11unorm',
-      'eac-r11snorm',
-      'eac-rg11unorm',
-      'eac-rg11snorm'
-    ])
-  ) {
-    basisFormats.push('etc2');
-  }
-  if (textureFormats.has('etc1-rgb-unorm-webgl')) {
-    basisFormats.push('etc1');
-  }
-  if (
-    hasSupportedTextureFormat(textureFormats, [
-      'atc-rgb-unorm-webgl',
-      'atc-rgba-unorm-webgl',
-      'atc-rgbai-unorm-webgl'
-    ])
-  ) {
-    basisFormats.push('atc-rgb', 'atc-rgba-interpolated-alpha');
-  }
-
-  basisFormats.push('rgba32', 'rgb565', 'bgr565', 'rgba4444');
-  return basisFormats;
-}
-
-function hasSupportedTextureFormat(
-  supportedTextureFormats: Set<TextureFormat>,
-  candidateTextureFormats: TextureFormat[]
-): boolean {
-  return candidateTextureFormats.some(textureFormat => supportedTextureFormats.has(textureFormat));
 }
